[refactor]: adjust module and class location. - move "entity_id" and "enhance_id" module to the same file, called "identity"
