+++ pydomain/basecls/__init__.py
@@ -1,5 +1,5 @@
-from .valueobj import ValueObject
-from .id import EntityId
+from .valueobject import ValueObject
+from .identity import EntityId, EntityEnhanceId
 from .entity import Entity
 from .aggregate import AggregateRoot
 from .errors import DomainErrorCode, DomainException
